+++ static/js/produccion/indexEquipos.js
@@ -1,7 +1,7 @@
 //Seccion equipos
 
 
-function list_equipo(url_list, url_edit, url_del) {
+function list_equipo(url_list, url_edit, url_del, perm_change, perm_delete) {
     $('#equipoTable').DataTable({
         responsive: true,
         pageLength: 6,
@@ -21,8 +21,13 @@ function list_equipo(url_list, url_edit, url_del) {
             targets: [-1],
             orderable: false,
             render: function (data, type, row) {
-                var buttons = '<button onclick="abrir_modal(\'' + url_edit + row.id + '/\')" class="btn btn-warning btn-sm" title="Editar"><i class="fas fa-edit"></i></button> ';
-                buttons += '<button onclick="delete_ajax_equipo(\'' + url_del + row.id + '/\')" title="Borrar" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i></button>';
+                let buttons = '';
+                if (perm_change){
+                    buttons += '<button onclick="abrir_modal(\'' + url_edit + row.id + '/\')" class="btn btn-warning btn-sm" title="Editar"><i class="fas fa-edit"></i></button> ';
+                }
+                if (perm_delete){
+                    buttons += '<button onclick="delete_ajax_equipo(\'' + url_del + row.id + '/\')" title="Borrar" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i></button>';
+                }
                 return buttons;
             }
         }]
